% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cross_validation.r
\name{cross_assess_wrapper}
\alias{cross_assess_wrapper}
\title{cross_valid_kfold}
\usage{
cross_assess_wrapper(data, formula, resp, family, K, model,
  cv_function = cross_valid_kfold)
}
\arguments{
\item{data, }{a data.frame containing response and observations variables used in the formula.}

\item{formula, }{relation of response ~ observation to be examined.}

\item{resp, }{a character vector of the names of dependent/response variables in df}

\item{family, }{a character string indicating the family associated with the submitted model c('gaussian','binomial','poisson'...)}

\item{K, }{a numeric variable indicating the number of folds for the k-fold validation}

\item{model, }{a model associated for testing the variables c(glm,lm)}

\item{cv_function, }{a crossvalidation function with the same inputs/outputs as cross_valid_kfold}
}
\value{
list of performance assessments
}
\description{
MAIN cross validation function. Includes performance assessment. Runs and assesses crossvalidation.
}

