% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cross_validation.r
\name{cross_valid_kfold}
\alias{cross_valid_kfold}
\title{cross_valid_kfold}
\usage{
cross_valid_kfold(data, formula, resp = all.vars(formula)[1], family,
  K = 10, model = glm)
}
\arguments{
\item{data, }{a data.frame containing response and observations variables used in the formula.}

\item{formula, }{relation of response ~ observation to be examined.}

\item{resp, }{a character vector of the names of dependent/response variables in df}

\item{family, }{a character string indicating the family associated with the submitted model c('gaussian','binomial','poisson'...)}

\item{K, }{a numeric variable indicating the number of folds for the k-fold validation}

\item{model, }{a model associated for testing the variables c(glm,lm)}
}
\value{
list of cross-validation results: pred_out matrix of predictions (observations x fold), train_out matrix of indicies used for each fold of cross-validation (observations x fold), labels_out matrix of ground truth labels (observations x fold)
}
\description{
k-fold cross validation run to collect predictions for assessment
}

