% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vis.r
\name{getGroupedMatrix}
\alias{getGroupedMatrix}
\title{getGroupedMatrix}
\usage{
getGroupedMatrix(m, k, groups)
}
\arguments{
\item{m}{matrix of coefficients}

\item{k}{number of prototype models to extract}

\item{groups}{group information (based on hierarchical tree)}
}
\value{
m raw coefficient matrix with grouped information

matL grouped coefficient matrix
}
\description{
getGroupedMatrix
}
