% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vis.r
\name{getReturnModels}
\alias{getReturnModels}
\title{getReturnModels}
\usage{
getReturnModels(l_reg, m, k)
}
\arguments{
\item{l_reg}{list of glmnet.cv objects}

\item{m}{matrix of coefficients}

\item{k}{number of prototype models to extract}
}
\value{
sel_return k models that are the centroids (most average) of each major cluster
}
\description{
getReturnModels
}
