% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vis.r
\name{getStatMatrix}
\alias{getStatMatrix}
\title{getStatMatrix}
\usage{
getStatMatrix(m, k, groups)
}
\arguments{
\item{m}{matrix of coefficients}

\item{k}{number of prototype models to extract}

\item{groups}{group information (based on hierarchical tree)}
}
\value{
m2 The matrix with average value of each group

m3 The matrix with variance of each group

m4 The matrix with F-score of each group
}
\description{
getStatMatrix
}
