% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vis.r
\name{loading_vis}
\alias{loading_vis}
\title{loading_vis}
\usage{
loading_vis(res.pca, var_catgories = NULL)
}
\arguments{
\item{res.pca}{a "prcomp" class object produced by PCA using prcomp}

\item{var_catgories}{an optional vector of categories into which each variable can be classified}
}
\value{
a ggplot2 object showing the loadings of the
}
\description{
Visualize loadings of a PCA
}
