% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RegressionModelPipeline.R
\name{model_selection}
\alias{model_selection}
\title{model_selection}
\usage{
model_selection(df, observations, response, family = "gaussian",
  model = glm, interactions = FALSE, test = c("Wald", "LRT"),
  thresh_screen = 0.2, only_return_selected = FALSE, K = 10,
  sig_vars_thresh = NULL, robust = FALSE, N = 1, aic_k = 2,
  robust_n = 100)
}
\arguments{
\item{df, }{a data.frame containing response and observations variables. Factors with more than 2 levels have only been implimented for test='LRT'}

\item{observations, }{a character vector of the names of independent/observations variables in df}

\item{response, }{a character vector of the names of dependent/response variables in df}

\item{family, }{a character string indicating the family associated with the submitted model c('gaussian','binomial','poisson'...)}

\item{model, }{a model associated for testing the variables c(glm,lm)}

\item{interactions, }{a boolean indicating if interactions should be assessed. Default is NULL, by default, interactions will be examined according to the constraints set by sig_vars_thresh. If a value is set for interactions (T/F) this will override the recomendations of sig_vars_thresh}

\item{test, }{a character string indicating Likelihood Ratio Test ('LRT') testing likelihood improvement of a model or Wald test ('Wald') testing coefficient > 0}

\item{thresh_screen, }{a numeric value indicating the p-value cutoff for the univariate screening}

\item{only_return_selected, }{a boolean value. If true, only models with p-value less than the threshold will be returned. Otherwise, all models will be returned.}

\item{K, }{a numeric value indicating the number of folds to use for k-fold cross-validation. K=10 by default. K=0 to skip k-fold validation.}

\item{sig_vars_thresh}{a list specifying the maximal number of significant variables allowed for each final model generating method. NULL (self initializing) by default.}

\item{robust}{boolean indicating if regularization will be run multiple times to get a robust indication of the underlying structure}

\item{N, }{a numeric value, default N=1, indicating the number of cross validation iterations to perform}

\item{robust_n, }{number of iterations for the robust glmnet run}
}
\value{
a list containing: univariate models, the final selected model, and crossvalidation stats.
}
\description{
Main function for model selection from a set of many variables
}
\examples{
mod=model_selection(df=mtcars,colnames(mtcars)[-1],response = 'mpg',interactions=F,test='LRT',K=5,family = 'gaussian',model=glm)
out=vis(mod)
print(out[[1]])
print(out[[2]])
}
