% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RegressionModelPipeline.R
\name{model_selection_many_to_few}
\alias{model_selection_many_to_few}
\title{model_selection_many_to_few}
\usage{
model_selection_many_to_few(...)
}
\arguments{
\item{..., }{see model_select parameters}
}
\value{
a list containing: univariate models, the final selected model, and crossvalidation stats.
}
\description{
runs regularization then model selection to reduce many variables to a small robust model TODO: Austin and Ben w/ vis_reg
}
