% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/performance_assessment.r
\name{pred_assess}
\alias{pred_assess}
\title{pred_assess}
\usage{
pred_assess(resp, pred = NULL, truth = NULL, mod = NULL, dat = NULL,
  ...)
}
\arguments{
\item{resp, }{a character vector of the names of dependent/response variables in df}

\item{pred, }{option 1 parameter: matrix of predictions (observations x fold)}

\item{truth, }{option 1 parameter: matrix of ground truth values (observations x fold)}

\item{mod, }{option 2 parameter: a trained model: glm(resp~...) or lm(resp~...)}

\item{dat, }{option 2 parameter: a data.frame containing all variables used in the mod model.}
}
\value{
list of performance assessments
}
\description{
MAIN cross validation assessment. There are 2 main usage modes. 1) provide predictions by setting pred and truth, 2) provide mod and dat to generate predictions to evaluate.
}
