\name{RegressionModelPipeline-package}
\alias{RegressionModelPipeline-package}
\alias{RegressionModelPipeline}
\docType{package}
\title{
\packageTitle{RegressionModelPipeline}
}
\description{
\packageDescription{RegressionModelPipeline}
}
\details{

The DESCRIPTION file:
\packageDESCRIPTION{RegressionModelPipeline}
\packageIndices{RegressionModelPipeline}
~~ An overview of how to use the package, including the most important functions ~~
}
\author{
\packageAuthor{RegressionModelPipeline}

Maintainer: \packageMaintainer{RegressionModelPipeline}
}
\references{
~~ Literature or other references for background information ~~
}
%~~ Optionally other standard keywords, one per line, from file KEYWORDS in the R documentation directory ~~
\keyword{ package }
\seealso{
~~ Optional links to other man pages, e.g. ~~
~~ \code{\link[<pkg>:<pkg>-package]{<pkg>}} ~~
}
\examples{
~~ simple examples of the most important functions ~~
}
