% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/multivariate_model.r
\name{stepwise_multivariate_model_selection}
\alias{stepwise_multivariate_model_selection}
\title{stepwise_multivariate_model_selection}
\usage{
stepwise_multivariate_model_selection(df, observations, response, family, model,
  interactions = FALSE)
}
\arguments{
\item{df, }{a data.frame containing response and observations variables. Factors with more than 2 levels have only been implimented for test='LRT'}

\item{observations, }{a character vector of the names of independent/observations variables in df}

\item{response, }{a character vector of the names of dependent/response variables in df}

\item{family, }{a character string indicating the family associated with the submitted model c('gaussian','binomial','poisson'...)}

\item{model, }{a model associated for testing the variables c(glm,lm)}

\item{interactions, }{a boolean indicating if interactions should be assessed. Default is False.}
}
\value{
a multivariate model
}
\description{
Multivariate model selection on all (significant) observations variables with respect to response.
}

