% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/univariate_screen.r
\name{univariate_screen}
\alias{univariate_screen}
\title{univariate_screen}
\usage{
univariate_screen(df, observations, response, family = "gaussian",
  model = glm, interactions = FALSE, test = "LRT", thresh = 0.2,
  only_return_selected = FALSE, context_var = NULL)
}
\arguments{
\item{df, }{a data.frame containing response and observations variables. Factors with more than 2 levels have only been implimented for test='LRT'}

\item{observations, }{a character vector of the names of independent/observations variables in df}

\item{response, }{a character vector of the names of dependent/response variables in df}

\item{family, }{a character string indicating the family associated with the submitted model c('gaussian','binomial','poisson'...)}

\item{model, }{a model associated for testing the variables c(glm,lm)}

\item{interactions, }{a boolean indicating if interactions should be assessed. Default is False.}

\item{test, }{a character string indicating Likelihood Ratio Test ('LRT') testing likelihood improvement of a model or Wald test ('Wald') testing coefficient > 0}

\item{thresh, }{a numeric value indicating the p-value cutoff for the univariate screening}

\item{only_return_selected, }{a boolean value. If true, only models with p-value less than the threshold will be returned. Otherwise, all models will be returned.}

\item{context_var, }{a character vector of variable(s) necessary for contextual assessment (e.g. age or time)}
}
\value{
a list of univariate models, attr(,'Pr') houses the significance of each model
}
\description{
Run univariate models on all observations variables with respect to response. Asseses p-value of model given LRT or Wald test. Returns all models or just models more significant than the given threshold: thresh.
}

