% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vis.r
\name{vis}
\alias{vis}
\title{vis}
\usage{
vis(out, family = "gaussian", Pr = NULL, fullUnivariate = FALSE,
  intercept = TRUE, col_scheme = rainbow(4), ...)
}
\arguments{
\item{out}{a length 2 list univariate and final models: list(screen=several_univariate_models,final=one_multivariate_model). Either screen or final may be set to NULL}

\item{family}{a character indicating the glm family fit for the model. If 'binomial' vis_logit will be run. Otherwise standard output is produced.}

\item{Pr}{an optional numeric vector of the same length and order as out$screen. If the user wants to specify univariate p-values they may do so here. Otherwise the wald test for the 1st variable in the model will be used.}

\item{fullUnivariate}{visualizes all variables in each univariate model}

\item{intercept}{boolean indicating if the intercept should be visualized. True by default.}

\item{col_scheme}{a grDevices::Palettes to specify the color scheme representing -log10(p-values). rainbow(4) by default.}
}
\value{
a list of 2 ggplot2 objects visualizing the univariate screening and multivariate final model.
}
\description{
A visualizaiton for univariate and multivariate regressions showing coefficient, confidence intervals and p-values
}

