% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vis.r
\name{vis_logit}
\alias{vis_logit}
\title{vis_logit}
\usage{
vis_logit(out, Pr = NULL, fullUnivariate = FALSE, intercept = TRUE,
  trans = "log2", col_scheme = rainbow(4))
}
\arguments{
\item{out}{a length 2 list univariate and final models: list(screen=several_univariate_models,final=one_multivariate_model). Either screen or final may be set to NULL}

\item{Pr}{an optional numeric vector of the same length and order as out$screen. If the user wants to specify univariate p-values they may do so here. Otherwise the wald test for the 1st variable in the model will be used.}

\item{fullUnivariate}{visualizes all variables in each univariate model}

\item{intercept}{boolean indicating if the intercept should be visualized. True by default.}

\item{trans}{a character indicating the transformation for the y-axis: log2, log10, sqrt. Refers to the trans parameter of scale_y_continuous. log2 by default.}

\item{col_scheme}{a grDevices::Palettes to specify the color scheme representing -log10(p-values). rainbow(4) by default.}
}
\value{
a list of 2 ggplot2 objects visualizing the univariate screening and multivariate final model.
}
\description{
A visualizaiton for logit regressions that calculates Odds Ratios (OR)
}

