% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vis.r
\name{vis_reg}
\alias{vis_reg}
\title{vis_reg}
\usage{
vis_reg(l_reg, k = 4)
}
\arguments{
\item{l_reg}{list of glmnet.cv objects}

\item{k}{number of prototype models to extract}
}
\value{
a list of the selected most prototypical models, 3 ggplot2 objects visualizing the coefficient, Average(coef.) and Sd(coef.) of l_reg models,
        and 3 matrix for plotting these 3 ggplot2 objects.
}
\description{
Visualize prototypical models for glmnets
}
